requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
fire==0.2.1
numpy==1.17.4             # via torch, torchvision
pillow==6.2.1             # via torchvision
six==1.13.0               # via fire, torchvision
termcolor==1.1.0          # via fire
torch==1.3.1
torchvision==0.4.2
